Align repo with new gitops flow cue structure
Our changes to the CUE repo structure as part of our GitOps flow design left
produce-aisle in a broken or otherwise inaccurate state. This work updated the
repo to bring it back to a working state: rewritting divergent CUE, removing
newly irrelevant files, etc.

Part of the new gitops flow design was the creation of a federated repository
strategy that includes a core repo governing critical mesh sections and then
lesser service repos maintained by developer teams. We decided to convert
produce-aisle (the long-standing de facto example for mesh configurations) into
the official "dev-team" service repo example. As a result, we tailored the
README to better inform the external user on what need to do.